Fixed linting

+++ common/fee/utils.go
@@ -8,7 +8,7 @@ import (
 )
 
 func ApplyMultiplier(feeLimit uint64, multiplier float32) (uint64, error) {
-	result := decimal.NewFromBigInt(big.NewInt(0).SetUint64(uint64(feeLimit)), 0).Mul(decimal.NewFromFloat32(multiplier)).BigInt()
+	result := decimal.NewFromBigInt(big.NewInt(0).SetUint64(feeLimit), 0).Mul(decimal.NewFromFloat32(multiplier)).BigInt()
 
 	if result.IsUint64() {
 		return 0, fmt.Errorf("integer cannot be represented as uint64 when applying multiplier of %f to fee limit of %d", multiplier, feeLimit)
